third_party/blink/web_tests/platform/linux-chrome/external/wpt/speculation-rules/prerender/prefetch.https-expected.txt

This is a testharness.js-based test.
[FAIL] Prerender navigation requests should use prefetched results
  assert_equals: expected "activated" but got "discarded"
Harness: the test ran to completion.
